Create read_multiple_tabs.py
I have an excel file that consists of 2 tabs: Bills and Invoices. The "Bill" tab contains 3 columns: BillID, Amount and Date. The "Invoices" tab contains 4 columns: InvoiceId, BillId, Amount and Date. Zero or more invoices may refer to the same bill and the sum of their amounts must be less or equal to the Bill's amount.
I need to read that excel and transform it to an xml message with the following structure:
<bills>
<bill>
<billId>123</billId>
<billAmount>123.10</billAmount>
<billDate>2020-12-10</billDate>
<invoices>
<invoice>
<invoiceId>123-001</invoiceId>
<invoiceAmount>100</invoiceAmount>
<invoiceDate>2020-12-13</invoiceDate>
</invoice>
<invoice>
<invoiceId>123-002</invoiceId>
<invoiceAmount>23.10</invoiceAmount>
<invoiceDate>2020-12-14</invoiceDate>
</invoice>
</bill>
<bill>
<billId>124</billId>
<billAmount>932.10</billAmount>
<billDate>2022-12-10</billDate>
<invoices/>
</bill>
</bills>